move winston setup to one file
